fix: Add debug logging, add householdreport as optional, cleanup of entities
