+++ surepetcare/entities/battery_mixin.py
