+++ surepetcare/security/exceptions.py
@@ -1,2 +1,8 @@
 class AuthenticationError(Exception):
     pass
+
+
+class ValidationMissingFieldsError(Exception):
+    """Raised when required fields are missing during model validation."""
+
+    pass
