+++ tests/devices/test_pet.py
@@ -2,11 +2,11 @@ import pytest
 
 from surepetcare.const import API_ENDPOINT_V2
 from surepetcare.devices.pet import Pet
-from surepetcare.entities.pet import ReportHouseholdDrinkingResource
-from surepetcare.entities.pet import ReportHouseholdFeedingResource
-from surepetcare.entities.pet import ReportHouseholdMovementResource
-from surepetcare.entities.pet import ReportHouseholdResource
-from surepetcare.entities.pet import ReportWeightFrame
+from surepetcare.devices.pet import ReportHouseholdDrinkingResource
+from surepetcare.devices.pet import ReportHouseholdFeedingResource
+from surepetcare.devices.pet import ReportHouseholdMovementResource
+from surepetcare.devices.pet import ReportHouseholdResource
+from surepetcare.devices.pet import ReportWeightFrame
 from tests.mock_helpers import MockSurePetcareClient
